+++ src/WebJobs.Script/Workers/Http/HttpFunctionInvocationDispatcher.cs
@@ -201,10 +201,5 @@ namespace Microsoft.Azure.WebJobs.Script.Workers
             await DisposeAndRestartWorkerChannel(_httpWorkerChannel.Id);    // Since there's only one channel for httpworker
             return true;
         }
-
-        public async Task RestartAllWorkersAsync()
-        {
-            await DisposeAndRestartWorkerChannel(_httpWorkerChannel.Id);    // Since there's only one channel for httpworker
-        }
     }
 }