+++ test/WebJobs.Script.Tests.Integration/TestFunctionHost.cs
@@ -119,6 +119,7 @@ namespace Microsoft.Azure.WebJobs.Script.Tests
             _testServer = new TestServer(builder) { BaseAddress = new Uri("https://localhost/") };
 
             HttpClient = _testServer.CreateClient();
+            HttpClient.Timeout = TimeSpan.FromMinutes(5);
 
             var manager = _testServer.Host.Services.GetService<IScriptHostManager>();
             _hostService = manager as WebJobsScriptHostService;
@@ -322,8 +323,7 @@ namespace Microsoft.Azure.WebJobs.Script.Tests
             var managerServiceProvider = manager as IServiceProvider;
 
             var metadataProvider = new FunctionMetadataProvider(optionsMonitor, NullLogger<FunctionMetadataProvider>.Instance, new TestMetricsLogger());
-            var metadataManager = new FunctionMetadataManager(managerServiceProvider.GetService<IOptions<ScriptJobHostOptions>>(),
-                metadataProvider, managerServiceProvider.GetService<IEnumerable<IFunctionProvider>>(),
+            var metadataManager = new FunctionMetadataManager(managerServiceProvider.GetService<IOptions<ScriptJobHostOptions>>(), metadataProvider,
                 managerServiceProvider.GetService<IOptions<HttpWorkerOptions>>(), manager, factory, new OptionsWrapper<LanguageWorkerOptions>(workerOptions));
 
             return metadataManager;